+++ udata_front/theme/gouvfr/assets/js/pages/Admin/Datasets/Datasets.vue
@@ -1,14 +1,14 @@
 <template>
-  <div class="fr-container--fluid">
+  <div>
     <Breadcrumb>
       <li>
         <router-link class="fr-breadcrumb__link" to="/">
           {{ t('Administration') }}
         </router-link>
       </li>
-      <li v-if="organization">
+      <li v-if="currentOrganization">
         <router-link class="fr-breadcrumb__link" to="/">
-          {{ organization.name }}
+          {{ currentOrganization.name }}
         </router-link>
       </li>
       <li>
@@ -17,31 +17,93 @@
         </a>
       </li>
     </Breadcrumb>
-    <div class="fr-alert fr-alert--info">
-      <h3 class="fr-alert__title">{{ t('This is a WIP page') }}</h3>
-      <p><a :href="url">{{ t('You can manage your datasets on the current admin.') }}</a></p>
+    <h1 class="fr-h3 fr-mb-5v">{{ t("Datasets") }}</h1>
+    <Container class="fr-mb-5v border">
+      <div class="fr-grid-row fr-grid-row--middle fr-grid-row--gutters">
+        <div class="fr-col">
+          <h2 class="fr-text--md fr-m-0">{{ t("Statistics of your datasets") }}</h2>
+        </div>
+        <p class="fr-col-auto fr-m-0">
+          <button class="fr-btn fr-btn--sm fr-icon-download-line fr-btn--icon-left">
+            {{ t("Download") }}
+          </button>
+        </p>
+        <!-- TODO: require useAccordion from @datagouv/components
+          <p class="fr-col-auto fr-ml-3v fr-m-0">
+            <button
+              @click="expand"
+              role="button"
+              :aria-expanded="expanded"
+              :aria-controls="resourceContentId"
+              class="fr-btn fr-btn--tertiary-no-outline fr-btn--icon-left fr-btn--secondary-grey-500"
+              :class="{'fr-icon-arrow-up-s-line': expanded, 'fr-icon-arrow-down-s-line': !expanded}"
+              data-testid="expand-button"
+            >
+            {{ t('See statistics') }}
+            </button>
+          </p>-->
+      </div>
+    </Container>
+    <div class="fr-grid-row fr-grid-row--gutters fr-grid-row--middle">
+      <h2 class="fr-col-12 fr-col-sm subtitle subtitle--uppercase fr-m-0">{{ t('{n} datasets', totalResult) }}</h2>
+      <div class="fr-col-auto fr-grid-row fr-grid-row--middle">
+        <!-- Add search component when org API allow search <InputGroup
+          class="fr-m-0 fr-mr-1w"
+          :label="t('Search')"
+          v-model.trim="q"
+        /> -->
+        <div>
+          <a :href="getOrganizationDatasetsCatalogUrl(oid)" class="fr-btn fr-btn--sm fr-icon-download-line fr-btn--icon-left">
+            {{ t('Download catalog') }}
+          </a>
+        </div>
+      </div>
     </div>
+    <DatasetsTable
+      class="fr-mt-1v"
+      :datasets="datasets"
+      @sort="(newSort) => sort = newSort"
+    />
+    <Pagination
+      v-if="totalResult > pageSize"
+      :page="page"
+      :page-size="pageSize"
+      :total-results="totalResult"
+      @change="(changedPage) => page = changedPage"
+    />
   </div>
 </template>
 <script setup lang="ts">
-import { onMounted, ref } from 'vue';
+import { Pagination, type Dataset } from "@datagouv/components";
+import { ref, watchEffect } from "vue";
 import { useI18n } from "vue-i18n";
+import { getOrganizationDatasets, getOrganizationDatasetsCatalogUrl } from "../../../api/datasets";
 import Breadcrumb from "../../../components/Breadcrumb/Breadcrumb.vue";
-import type { Me } from "../../../types";
-import { fetchMe } from "../../../api/me";
-import { watchEffect } from 'vue';
-import type { Organization } from '@datagouv/components';
+import DatasetsTable from "../../../components/AdminTable/AdminDatasetsTable/AdminDatasetsTable.vue";
+import Container from "../../../components/Ui/Container/Container.vue";
+import InputGroup from "../../../components/Form/InputGroup/InputGroup.vue";
+import { useCurrentOrganization } from '../../../composables/admin/useCurrentOrganization';
+import { refDebounced } from "@vueuse/core";
+import { search_autocomplete_debounce } from "../../../config";
 
 const { t } = useI18n();
 const props = defineProps<{oid: string}>();
-const url = `/admin/organization/${props.oid}/`;
 
-const me = ref<Me | null>(null);
-const organization = ref<Organization | null>(null);
+const datasets = ref<Array<Dataset>>([]);
+const page = ref(1);
+const pageSize = ref(10);
+const totalResult = ref(0);
+const sort = ref("-created");
+const q = ref("");
+const qDebounced = refDebounced(q, search_autocomplete_debounce);
 
-onMounted(async () => me.value = await fetchMe());
+const { currentOrganization } = useCurrentOrganization();
 
-watchEffect(() => {
-  organization.value = me.value?.organizations.find(organization => organization.id === props.oid) ?? null;
+watchEffect(async () => {
+  const response = await getOrganizationDatasets(props.oid, qDebounced.value, page.value, pageSize.value, sort.value);
+  datasets.value = response.data;
+  page.value = response.page;
+  pageSize.value = response.page_size;
+  totalResult.value = response.total;
 });
 </script>