+++ udata_front/theme/gouvfr/assets/js/pages/Admin/Members/Members.vue
@@ -12,6 +12,9 @@
         </router-link>
       </li>
       <li>
+        <a class="fr-breadcrumb__link" aria-current="page">
+          {{ t('Members') }}
+        </a>
       </li>
     </Breadcrumb>
     <div class="fr-grid-row fr-grid-row--gutters fr-grid-row--middle justify-center">
@@ -81,12 +84,13 @@ import { useI18n } from "vue-i18n";
 import { acceptRequest, formatRolesAsOptions, getOrganization, getPendingMemberships, getRoles, refuseRequest } from "../../../api/organizations";
 import Breadcrumb from "../../../components/Breadcrumb/Breadcrumb.vue";
 import { type Option } from "../../../components/Form/SelectGroup/SelectGroup.vue";
+import AdminAddMemberButton from "../../../components/AdminAddMember/AdminAddMemberButton.vue";
 import AdminEditMemberButton from "../../../components/AdminEditMember/AdminEditMemberButton.vue";
 import AdminMembershipRequest from "../../../components/AdminMembershipRequest/AdminMembershipRequest.vue";
-import type { EditingMember, MemberRole, PendingMembershipRequest } from "../../../types";
+import { useCurrentOrganization } from "../../../composables/admin/useCurrentOrganization";
 import { useToast } from "../../../composables/useToast";
 import { user, userIsAdmin } from "../../../config";
-import AdminAddMemberButton from "../../../components/AdminAddMember/AdminAddMemberButton.vue";
+import type { EditingMember, MemberRole, PendingMembershipRequest } from "../../../types";
 
 const props = defineProps<{oid: string}>();
 
@@ -100,8 +104,7 @@ const isAdmin = computed(() => userIsAdmin || members.value.some(member => membe
 
 const roles = ref<Array<Option>>([]);
 
-// TODO : use `useCurrentOrganization` when merged
-const currentOrganization = {name: "SomeName"};
+const { currentOrganization } = useCurrentOrganization();
 
 const members = ref<Array<EditingMember>>([]);